SVN: merge trunk-3-4 to trunk-3-5
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk-3-5@13727 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/graph/runtime/jmx/GraphTrackingDetail.java
@@ -166,7 +166,7 @@ public class GraphTrackingDetail implements GraphTracking {
 	}
 
 	@Override
-	public int getUsedMemory() {
+	public long getUsedMemory() {
 		return graph.getMemoryTracker().getUsedMemory();
 	}
